Move all common test data into a common file.
