+++ graphql_compiler/tests/test_compiler.py
@@ -6,36 +6,33 @@ from graphql import GraphQLID, GraphQLInt, GraphQLList, GraphQLString
 import six
 
 from ..compiler import OutputMetadata, compile_graphql_to_gremlin, compile_graphql_to_match
-from ..exceptions import GraphQLCompilationError
-from ..schema import GraphQLDate, GraphQLDateTime
 from .test_helpers import compare_gremlin, compare_input_metadata, compare_match, get_schema
 from . import test_input_data
 
 
-def check_test_data(test_case, graphql_input, expected_match, expected_gremlin,
-                    expected_output_metadata, expected_input_metadata, type_equivalence_hints=None):
+def check_test_data(test_case, test_data, expected_match, expected_gremlin):
     """Assert that the GraphQL input generates all expected MATCH and Gremlin data."""
-    if type_equivalence_hints:
+    if test_data.type_equivalence_hints:
         # For test convenience, we accept the type equivalence hints in string form.
         # Here, we convert them to the required GraphQL types.
         schema_based_type_equivalence_hints = {
             test_case.schema.get_type(key): test_case.schema.get_type(value)
-            for key, value in six.iteritems(type_equivalence_hints)
+            for key, value in six.iteritems(test_data.type_equivalence_hints)
         }
     else:
         schema_based_type_equivalence_hints = None
 
-    result = compile_graphql_to_match(test_case.schema, graphql_input,
+    result = compile_graphql_to_match(test_case.schema, test_data.graphql_input,
                                       type_equivalence_hints=schema_based_type_equivalence_hints)
     compare_match(test_case, expected_match, result.query)
-    test_case.assertEqual(expected_output_metadata, result.output_metadata)
-    compare_input_metadata(test_case, expected_input_metadata, result.input_metadata)
+    test_case.assertEqual(test_data.expected_output_metadata, result.output_metadata)
+    compare_input_metadata(test_case, test_data.expected_input_metadata, result.input_metadata)
 
-    result = compile_graphql_to_gremlin(test_case.schema, graphql_input,
+    result = compile_graphql_to_gremlin(test_case.schema, test_data.graphql_input,
                                         type_equivalence_hints=schema_based_type_equivalence_hints)
     compare_gremlin(test_case, expected_gremlin, result.query)
-    test_case.assertEqual(expected_output_metadata, result.output_metadata)
-    compare_input_metadata(test_case, expected_input_metadata, result.input_metadata)
+    test_case.assertEqual(test_data.expected_output_metadata, result.output_metadata)
+    compare_input_metadata(test_case, test_data.expected_input_metadata, result.input_metadata)
 
 
 class CompilerTests(unittest.TestCase):
@@ -64,8 +61,7 @@ class CompilerTests(unittest.TestCase):
             ])}
         '''
 
-        check_test_data(self, test_data.graphql_input, expected_match, expected_gremlin,
-                        test_data.expected_output_metadata, test_data.expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_immediate_filter_and_output(self):
         # Ensure that all basic comparison operators output correct code in this simple case.
@@ -108,17 +104,16 @@ class CompilerTests(unittest.TestCase):
                 'wanted': GraphQLString,
             }
 
-            check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                            expected_output_metadata, expected_input_metadata)
+            test_data = test_input_data.CommonTestData(
+                graphql_input=graphql_input,
+                expected_output_metadata=expected_output_metadata,
+                expected_input_metadata=expected_input_metadata,
+                type_equivalence_hints=None)
+
+            check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_multiple_filters(self):
-        graphql_input = '''{
-            Animal {
-                name @filter(op_name: ">=", value: ["$lower_bound"])
-                     @filter(op_name: "<", value: ["$upper_bound"])
-                     @output(out_name: "animal_name")
-            }
-        }'''
+        test_data = test_input_data.multiple_filters()
 
         expected_match = '''
             SELECT Animal___1.name AS `animal_name` FROM (
@@ -139,25 +134,11 @@ class CompilerTests(unittest.TestCase):
                 animal_name: m.Animal___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'lower_bound': GraphQLString,
-            'upper_bound': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_traverse_and_output(self):
-        graphql_input = '''{
-            Animal {
-                out_Animal_ParentOf {
-                    name @output(out_name: "parent_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.traverse_and_output()
 
         expected_match = '''
             SELECT Animal__out_Animal_ParentOf___1.name AS `parent_name` FROM (
@@ -180,25 +161,11 @@ class CompilerTests(unittest.TestCase):
                 parent_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'parent_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_optional_traverse_after_mandatory_traverse(self):
-        graphql_input = '''{
-            Animal {
-                out_Animal_OfSpecies {
-                    name @output(out_name: "species_name")
-                }
-                out_Animal_ParentOf @optional {
-                    name @output(out_name: "child_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.optional_traverse_after_mandatory_traverse()
 
         expected_match = '''
             SELECT if(eval("(Animal__out_Animal_ParentOf___1 IS NOT null)"),
@@ -236,23 +203,11 @@ class CompilerTests(unittest.TestCase):
                 species_name: m.Animal__out_Animal_OfSpecies___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'child_name': OutputMetadata(type=GraphQLString, optional=True),
-            'species_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_traverse_filter_and_output(self):
-        graphql_input = '''{
-            Animal {
-                out_Animal_ParentOf @filter(op_name: "name_or_alias", value: ["$wanted"]) {
-                    name @output(out_name: "parent_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.traverse_filter_and_output()
 
         expected_match = '''
             SELECT Animal__out_Animal_ParentOf___1.name AS `parent_name` FROM (
@@ -277,24 +232,11 @@ class CompilerTests(unittest.TestCase):
                 parent_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'parent_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'wanted': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_name_or_alias_filter_on_interface_type(self):
-        graphql_input = '''{
-            Animal {
-                out_Entity_Related @filter(op_name: "name_or_alias", value: ["$wanted"]) {
-                    name @output(out_name: "related_entity")
-                }
-            }
-        }'''
+        test_data = test_input_data.name_or_alias_filter_on_interface_type()
 
         expected_match = '''
             SELECT Animal__out_Entity_Related___1.name AS `related_entity` FROM (
@@ -319,25 +261,11 @@ class CompilerTests(unittest.TestCase):
                 related_entity: m.Animal__out_Entity_Related___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'related_entity': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'wanted': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_output_source_and_complex_output(self):
-        graphql_input = '''{
-            Animal {
-                name @filter(op_name: "=", value: ["$wanted"]) @output(out_name: "animal_name")
-                out_Animal_ParentOf @output_source {
-                    name @output(out_name: "parent_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.output_source_and_complex_output()
 
         expected_match = '''
             SELECT
@@ -365,32 +293,12 @@ class CompilerTests(unittest.TestCase):
                 parent_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'parent_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'wanted': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_on_optional_variable_equality(self):
         # The operand in the @filter directive originates from an optional block.
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ParentOf {
-                    out_Animal_FedAt @optional {
-                        name @tag(tag_name: "child_fed_at_event")
-                    }
-                }
-                out_Animal_FedAt @output_source {
-                    name @filter(op_name: "=", value: ["%child_fed_at_event"])
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_on_optional_variable_equality()
 
         expected_match = '''
 SELECT
@@ -437,27 +345,12 @@ FROM (
                 animal_name: m.Animal___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_on_optional_variable_name_or_alias(self):
         # The operand in the @filter directive originates from an optional block.
-        graphql_input = '''{
-            Animal {
-                in_Animal_ParentOf @optional {
-                    name @tag(tag_name: "parent_name")
-                }
-                out_Animal_ParentOf @filter(op_name: "name_or_alias", value: ["%parent_name"])
-                                    @output_source {
-                    name @output(out_name: "animal_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_on_optional_variable_name_or_alias()
 
         expected_match = '''
 SELECT
@@ -505,23 +398,11 @@ FROM (
                 animal_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_in_optional_block(self):
-        graphql_input = '''{
-            Animal {
-                out_Animal_FedAt @optional {
-                    name @filter(op_name: "=", value: ["$name"])
-                    uuid @output(out_name: "uuid")
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_in_optional_block()
 
         expected_match = '''
             SELECT
@@ -552,48 +433,13 @@ FROM (
                        m.Animal__out_Animal_FedAt___1.uuid : null)
             ])}
         '''
-        expected_output_metadata = {
-            'uuid': OutputMetadata(type=GraphQLID, optional=True),
-        }
-        expected_input_metadata = {
-            'name': GraphQLString
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
-
-    def test_optional_traversal_edge_case(self):
-        # Both Animal and out_Animal_ParentOf have an out_Animal_FedAt field,
-        # ensure the correct such field is picked out.
-        graphql_input = '''{
-            Animal {
-                out_Animal_ParentOf @optional {
-                    out_Animal_FedAt {
-                        name @output(out_name: "name")
-                    }
-                }
-            }
-        }'''
-
-        # Disabled until OrientDB fixes the limitation against traversing from an optional vertex.
-        # Rather than compiling correctly, this raises an exception until the OrientDB limitation
-        # is lifted.
-        # For details, see https://github.com/orientechnologies/orientdb/issues/6788
-        with self.assertRaises(GraphQLCompilationError):
-            compile_graphql_to_match(self.schema, graphql_input)
-
-        with self.assertRaises(GraphQLCompilationError):
-            compile_graphql_to_gremlin(self.schema, graphql_input)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_between_filter_on_simple_scalar(self):
         # The "between" filter emits different output depending on what the compared types are.
         # This test checks for correct code generation when the type is a simple scalar (a String).
-        graphql_input = '''{
-            Animal {
-                name @filter(op_name: "between", value: ["$lower", "$upper"])
-                     @output(out_name: "name")
-            }
-        }'''
+        test_data = test_input_data.between_filter_on_simple_scalar()
 
         expected_match = '''
             SELECT
@@ -615,26 +461,13 @@ FROM (
                 name: m.Animal___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'lower': GraphQLString,
-            'upper': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_between_filter_on_date(self):
         # The "between" filter emits different output depending on what the compared types are.
         # This test checks for correct code generation when the type is a custom scalar (Date).
-        graphql_input = '''{
-            Animal {
-                birthday @filter(op_name: "between", value: ["$lower", "$upper"])
-                         @output(out_name: "birthday")
-            }
-        }'''
+        test_data = test_input_data.between_filter_on_date()
 
         expected_match = '''
             SELECT
@@ -662,26 +495,13 @@ FROM (
                 birthday: m.Animal___1.birthday.format("yyyy-MM-dd")
             ])}
         '''
-        expected_output_metadata = {
-            'birthday': OutputMetadata(type=GraphQLDate, optional=False),
-        }
-        expected_input_metadata = {
-            'lower': GraphQLDate,
-            'upper': GraphQLDate,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_between_filter_on_datetime(self):
         # The "between" filter emits different output depending on what the compared types are.
         # This test checks for correct code generation when the type is a custom scalar (DateTime).
-        graphql_input = '''{
-            Event {
-                event_date @filter(op_name: "between", value: ["$lower", "$upper"])
-                           @output(out_name: "event_date")
-            }
-        }'''
+        test_data = test_input_data.between_filter_on_datetime()
 
         expected_match = '''
             SELECT
@@ -709,46 +529,13 @@ FROM (
                 event_date: m.Event___1.event_date.format("yyyy-MM-dd'T'HH:mm:ssX")
             ])}
         '''
-        expected_output_metadata = {
-            'event_date': OutputMetadata(type=GraphQLDateTime, optional=False),
-        }
-        expected_input_metadata = {
-            'lower': GraphQLDateTime,
-            'upper': GraphQLDateTime,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_complex_optional_variables(self):
         # The operands in the @filter directives originate from an optional block,
         # in addition to having very complex filtering logic.
-        graphql_input = '''{
-            Animal {
-                name @filter(op_name: "=", value: ["$animal_name"])
-                out_Animal_ParentOf {
-                    out_Animal_FedAt @optional {
-                        name @tag(tag_name: "child_fed_at_event")
-                        event_date @tag(tag_name: "child_fed_at")
-                                   @output(out_name: "child_fed_at")
-                    }
-                    in_Animal_ParentOf {
-                        out_Animal_FedAt @optional {
-                            event_date @tag(tag_name: "other_parent_fed_at")
-                                       @output(out_name: "other_parent_fed_at")
-                        }
-                    }
-                }
-                in_Animal_ParentOf {
-                    out_Animal_FedAt {
-                        name @filter(op_name: "=", value: ["%child_fed_at_event"])
-                        event_date @output(out_name: "grandparent_fed_at")
-                                   @filter(op_name: "between",
-                                           value: ["%other_parent_fed_at", "%child_fed_at"])
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.complex_optional_variables()
 
         expected_match = '''
 SELECT
@@ -864,32 +651,11 @@ g.V('@class', 'Animal')
     )
 ])}
         '''
-        expected_output_metadata = {
-            'child_fed_at': OutputMetadata(type=GraphQLDateTime, optional=True),
-            'other_parent_fed_at': OutputMetadata(type=GraphQLDateTime, optional=True),
-            'grandparent_fed_at': OutputMetadata(type=GraphQLDateTime, optional=False),
-        }
-        expected_input_metadata = {
-            'animal_name': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_simple_fragment(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Entity_Related {
-                    ... on Animal {
-                        name @output(out_name: "related_animal_name")
-                        out_Animal_OfSpecies {
-                            name @output(out_name: "related_animal_species")
-                        }
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.simple_fragment()
 
         expected_match = '''
 SELECT
@@ -925,25 +691,11 @@ FROM (
                 related_animal_species: m.Animal__out_Entity_Related__out_Animal_OfSpecies___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'related_animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'related_animal_species': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_typename_output(self):
-        graphql_input = '''{
-            Animal {
-                __typename @output(out_name: "base_cls")
-                out_Animal_OfSpecies {
-                    __typename @output(out_name: "child_cls")
-                }
-            }
-        }'''
+        test_data = test_input_data.typename_output()
 
         expected_match = '''
             SELECT
@@ -970,22 +722,11 @@ FROM (
                 child_cls: m.Animal__out_Animal_OfSpecies___1['@class']
             ])}
         '''
-        expected_output_metadata = {
-            'base_cls': OutputMetadata(type=GraphQLString, optional=False),
-            'child_cls': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_typename_filter(self):
-        graphql_input = '''{
-            Entity {
-                __typename @filter(op_name: "=", value: ["$base_cls"])
-                name @output(out_name: "entity_name")
-            }
-        }'''
+        test_data = test_input_data.typename_filter()
 
         expected_match = '''
             SELECT
@@ -1007,24 +748,11 @@ FROM (
                 entity_name: m.Entity___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'entity_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'base_cls': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_simple_recurse(self):
-        graphql_input = '''{
-            Animal {
-                out_Animal_ParentOf @recurse(depth: 1) {
-                    name @output(out_name: "relation_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.simple_recurse()
 
         expected_match = '''
             SELECT
@@ -1054,28 +782,11 @@ FROM (
                 relation_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'relation_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_recurse_within_fragment(self):
-        graphql_input = '''{
-            Food {
-                name @output(out_name: "food_name")
-                in_Entity_Related {
-                    ... on Animal {
-                        name @output(out_name: "animal_name")
-                        out_Animal_ParentOf @recurse(depth: 3) {
-                            name @output(out_name: "relation_name")
-                        }
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.recurse_within_fragment()
 
         expected_match = '''
             SELECT
@@ -1118,25 +829,11 @@ FROM (
                 relation_name: m.Food__in_Entity_Related__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'food_name': OutputMetadata(type=GraphQLString, optional=False),
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'relation_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_within_recurse(self):
-        graphql_input = '''{
-            Animal {
-                out_Animal_ParentOf @recurse(depth: 3) {
-                    name @output(out_name: "relation_name")
-                    color @filter(op_name: "=", value: ["$wanted"])
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_within_recurse()
 
         expected_match = '''
             SELECT
@@ -1170,26 +867,11 @@ FROM (
                 relation_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'relation_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'wanted': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_recurse_with_immediate_type_coercion(self):
-        graphql_input = '''{
-            Animal {
-                in_Entity_Related @recurse(depth: 4) {
-                    ... on Animal {
-                        name @output(out_name: "name")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.recurse_with_immediate_type_coercion()
 
         expected_match = '''
             SELECT
@@ -1225,25 +907,11 @@ FROM (
                 name: m.Animal__in_Entity_Related___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_recurse_with_immediate_type_coercion_and_filter(self):
-        graphql_input = '''{
-            Animal {
-                in_Entity_Related @recurse(depth: 4) {
-                    ... on Animal {
-                        name @output(out_name: "name")
-                        color @filter(op_name: "=", value: ["$color"])
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.recurse_with_immediate_type_coercion_and_filter()
 
         expected_match = '''
             SELECT
@@ -1280,23 +948,11 @@ FROM (
                 name: m.Animal__in_Entity_Related___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'color': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_in_collection_op_filter_with_variable(self):
-        graphql_input = '''{
-            Animal {
-                name @filter(op_name: "in_collection", value: ["$wanted"])
-                     @output(out_name: "animal_name")
-            }
-        }'''
+        test_data = test_input_data.in_collection_op_filter_with_variable()
 
         expected_match = '''
             SELECT
@@ -1318,26 +974,11 @@ FROM (
                 animal_name: m.Animal___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'wanted': GraphQLList(GraphQLString),
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_in_collection_op_filter_with_tag(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                alias @tag(tag_name: "aliases")
-                out_Animal_ParentOf {
-                    name @filter(op_name: "in_collection", value: ["%aliases"])
-                }
-            }
-        }'''
+        test_data = test_input_data.in_collection_op_filter_with_tag()
 
         expected_match = '''
             SELECT
@@ -1364,26 +1005,11 @@ FROM (
                 animal_name: m.Animal___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_in_collection_op_filter_with_optional_tag(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                in_Animal_ParentOf @optional {
-                    alias @tag(tag_name: "parent_aliases")
-                }
-                out_Animal_ParentOf {
-                    name @filter(op_name: "in_collection", value: ["%parent_aliases"])
-                }
-            }
-        }'''
+        test_data = test_input_data.in_collection_op_filter_with_optional_tag()
 
         expected_match = '''
             SELECT
@@ -1426,13 +1052,8 @@ FROM (
                 animal_name: m.Animal___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_has_substring_op_filter_with_variable(self):
         graphql_input = '''{
@@ -1469,8 +1090,13 @@ FROM (
             'wanted': GraphQLString,
         }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        test_data = test_input_data.CommonTestData(
+            graphql_input=graphql_input,
+            expected_output_metadata=expected_output_metadata,
+            expected_input_metadata=expected_input_metadata,
+            type_equivalence_hints=None)
+
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_has_substring_op_filter_with_tag(self):
         graphql_input = '''{
@@ -1512,8 +1138,13 @@ FROM (
         }
         expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        test_data = test_input_data.CommonTestData(
+            graphql_input=graphql_input,
+            expected_output_metadata=expected_output_metadata,
+            expected_input_metadata=expected_input_metadata,
+            type_equivalence_hints=None)
+
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_has_substring_op_filter_with_optional_tag(self):
         graphql_input = '''{
@@ -1574,19 +1205,16 @@ FROM (
         }
         expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        test_data = test_input_data.CommonTestData(
+            graphql_input=graphql_input,
+            expected_output_metadata=expected_output_metadata,
+            expected_input_metadata=expected_input_metadata,
+            type_equivalence_hints=None)
+
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_has_edge_degree_op_filter(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ParentOf @filter(op_name: "has_edge_degree", value: ["$child_count"])
-                                    @output_source {
-                    name @output(out_name: "child_name")
-                }
-            }
-        }'''
+        test_data = test_input_data.has_edge_degree_op_filter()
 
         expected_match = '''
             SELECT
@@ -1622,32 +1250,11 @@ FROM (
                 child_name: m.Animal__out_Animal_ParentOf___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'child_count': GraphQLInt,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_has_edge_degree_op_filter_with_optional(self):
-        graphql_input = '''{
-            Species {
-                name @output(out_name: "species_name")
-
-                in_Animal_OfSpecies {
-                    name @output(out_name: "parent_name")
-
-                    out_Animal_ParentOf @filter(op_name: "has_edge_degree", value: ["$child_count"])
-                                        @optional {
-                        name @output(out_name: "child_name")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.has_edge_degree_op_filter_with_optional()
 
         expected_match = '''
             SELECT
@@ -1698,33 +1305,11 @@ FROM (
                 species_name: m.Species___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'species_name': OutputMetadata(type=GraphQLString, optional=False),
-            'parent_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_name': OutputMetadata(type=GraphQLString, optional=True),
-        }
-        expected_input_metadata = {
-            'child_count': GraphQLInt,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_has_edge_degree_op_filter_with_fold(self):
-        graphql_input = '''{
-            Species {
-                name @output(out_name: "species_name")
-
-                in_Animal_OfSpecies {
-                    name @output(out_name: "parent_name")
-
-                    out_Animal_ParentOf @filter(op_name: "has_edge_degree", value: ["$child_count"])
-                                        @fold {
-                        name @output(out_name: "child_names")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.has_edge_degree_op_filter_with_fold()
 
         expected_match = '''
             SELECT
@@ -1769,29 +1354,11 @@ FROM (
                 species_name: m.Species___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'species_name': OutputMetadata(type=GraphQLString, optional=False),
-            'parent_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_names': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {
-            'child_count': GraphQLInt,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_simple_union(self):
-        graphql_input = '''{
-            Species {
-                name @output(out_name: "species_name")
-                out_Species_Eats {
-                    ... on Food {
-                        name @output(out_name: "food_name")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.simple_union()
 
         expected_match = '''
             SELECT
@@ -1820,26 +1387,11 @@ FROM (
                 species_name: m.Species___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'food_name': OutputMetadata(type=GraphQLString, optional=False),
-            'species_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_on_fragment_in_union(self):
-        graphql_input = '''{
-            Species {
-                name @output(out_name: "species_name")
-                out_Species_Eats {
-                    ... on Food @filter(op_name: "name_or_alias", value: ["$wanted"]) {
-                        name @output(out_name: "food_name")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_on_fragment_in_union()
 
         expected_match = '''
             SELECT
@@ -1870,28 +1422,11 @@ FROM (
                 species_name: m.Species___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'food_name': OutputMetadata(type=GraphQLString, optional=False),
-            'species_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {
-            'wanted': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_optional_on_union(self):
-        graphql_input = '''{
-            Species {
-                name @output(out_name: "species_name")
-                out_Species_Eats @optional {
-                    ... on Food {
-                        name @output(out_name: "food_name")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.optional_on_union()
 
         expected_match = '''
             SELECT
@@ -1926,14 +1461,8 @@ FROM (
                 species_name: m.Species___1.name
             ])}
         '''
-        expected_output_metadata = {
-            'food_name': OutputMetadata(type=GraphQLString, optional=True),
-            'species_name': OutputMetadata(type=GraphQLString, optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_gremlin_type_hints(self):
         graphql_input = '''{
@@ -1979,9 +1508,13 @@ FROM (
         }
         expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata,
-                        type_equivalence_hints=type_equivalence_hints)
+        test_data = test_input_data.CommonTestData(
+            graphql_input=graphql_input,
+            expected_output_metadata=expected_output_metadata,
+            expected_input_metadata=expected_input_metadata,
+            type_equivalence_hints=type_equivalence_hints)
+
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_unnecessary_traversal_elimination(self):
         # This test case caught a bug in the optimization pass that eliminates unnecessary
@@ -2079,18 +1612,16 @@ FROM (
             'uuid': GraphQLID,
         }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        test_data = test_input_data.CommonTestData(
+            graphql_input=graphql_input,
+            expected_output_metadata=expected_output_metadata,
+            expected_input_metadata=expected_input_metadata,
+            type_equivalence_hints=None)
+
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_fold_on_output_variable(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ParentOf @fold {
-                    name @output(out_name: "child_names_list")
-                }
-            }
-        }'''
+        test_data = test_input_data.fold_on_output_variable()
 
         expected_match = '''
             SELECT
@@ -2119,26 +1650,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_names_list': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_fold_after_traverse(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                in_Animal_ParentOf {
-                    out_Animal_ParentOf @fold {
-                        name @output(out_name: "sibling_and_self_names_list")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.fold_after_traverse()
 
         expected_match = '''
             SELECT
@@ -2175,26 +1691,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'sibling_and_self_names_list': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_multiple_outputs_in_same_fold(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ParentOf @fold {
-                    name @output(out_name: "child_names_list")
-                    uuid @output(out_name: "child_uuids_list")
-                }
-            }
-        }'''
+        test_data = test_input_data.multiple_outputs_in_same_fold()
 
         expected_match = '''
             SELECT
@@ -2228,30 +1729,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_names_list': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-            'child_uuids_list': OutputMetadata(type=GraphQLList(GraphQLID), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_multiple_folds(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ParentOf @fold {
-                    name @output(out_name: "child_names_list")
-                    uuid @output(out_name: "child_uuids_list")
-                }
-                in_Animal_ParentOf @fold {
-                    name @output(out_name: "parent_names_list")
-                    uuid @output(out_name: "parent_uuids_list")
-                }
-            }
-        }'''
+        test_data = test_input_data.multiple_folds()
 
         expected_match = '''
             SELECT
@@ -2298,30 +1780,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_names_list': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-            'child_uuids_list': OutputMetadata(type=GraphQLList(GraphQLID), optional=False),
-            'parent_names_list': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-            'parent_uuids_list': OutputMetadata(type=GraphQLList(GraphQLID), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_fold_date_and_datetime_fields(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ParentOf @fold {
-                    birthday @output(out_name: "child_birthdays_list")
-                }
-                out_Animal_FedAt @fold {
-                    event_date @output(out_name: "fed_at_datetimes_list")
-                }
-            }
-        }'''
+        test_data = test_input_data.fold_date_and_datetime_fields()
 
         expected_match = '''
             SELECT
@@ -2362,30 +1825,14 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_birthdays_list': OutputMetadata(type=GraphQLList(GraphQLDate), optional=False),
-            'fed_at_datetimes_list': OutputMetadata(
-                type=GraphQLList(GraphQLDateTime), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_coercion_to_union_base_type_inside_fold(self):
         # Given type_equivalence_hints = { Event: EventOrBirthEvent },
         # the coercion should be optimized away as a no-op.
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Animal_ImportantEvent @fold {
-                    ... on Event {
-                        name @output(out_name: "important_events")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.coercion_to_union_base_type_inside_fold()
+
         type_equivalence_hints = {
             'Event': 'EventOrBirthEvent'
         }
@@ -2419,28 +1866,12 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'important_events': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata,
-                        type_equivalence_hints=type_equivalence_hints)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_no_op_coercion_inside_fold(self):
         # The type where the coercion is applied is already Entity, so the coercion is a no-op.
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "animal_name")
-                out_Entity_Related @fold {
-                    ... on Entity {
-                        name @output(out_name: "related_entities")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.no_op_coercion_inside_fold()
 
         expected_match = '''
             SELECT
@@ -2470,25 +1901,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'animal_name': OutputMetadata(type=GraphQLString, optional=False),
-            'related_entities': OutputMetadata(type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_within_fold_scope(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "name")
-                out_Animal_ParentOf @fold {
-                    name @filter(op_name: "=", value: ["$desired"]) @output(out_name: "child_list")
-                    description @output(out_name: "child_descriptions")
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_within_fold_scope()
 
         expected_match = '''
             SELECT
@@ -2529,30 +1946,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_list': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-            'child_descriptions': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {
-            'desired': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_filter_on_fold_scope(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "name")
-                out_Animal_ParentOf @fold
-                                    @filter(op_name: "name_or_alias", value: ["$desired"]) {
-                    name @output(out_name: "child_list")
-                }
-            }
-        }'''
+        test_data = test_input_data.filter_on_fold_scope()
 
         expected_match = '''
             SELECT
@@ -2586,29 +1984,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-            'child_list': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {
-            'desired': GraphQLString,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_coercion_on_interface_within_fold_scope(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "name")
-                out_Entity_Related @fold {
-                    ... on Animal {
-                        name @output(out_name: "related_animals")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.coercion_on_interface_within_fold_scope()
 
         expected_match = '''
             SELECT
@@ -2640,27 +2020,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-            'related_animals': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_coercion_on_union_within_fold_scope(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "name")
-                out_Animal_ImportantEvent @fold {
-                    ... on BirthEvent {
-                        name @output(out_name: "birth_events")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.coercion_on_union_within_fold_scope()
 
         expected_match = '''
             SELECT
@@ -2692,30 +2056,11 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-            'birth_events': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-        }
-        expected_input_metadata = {}
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
 
     def test_coercion_filters_and_multiple_outputs_within_fold_scope(self):
-        graphql_input = '''{
-            Animal {
-                name @output(out_name: "name")
-                out_Entity_Related @fold {
-                    ... on Animal {
-                        name @filter(op_name: "has_substring", value: ["$substring"])
-                             @output(out_name: "related_animals")
-                        birthday @filter(op_name: "<=", value: ["$latest"])
-                                 @output(out_name: "related_birthdays")
-                    }
-                }
-            }
-        }'''
+        test_data = test_input_data.coercion_filters_and_multiple_outputs_within_fold_scope()
 
         expected_match = '''
     SELECT
@@ -2765,17 +2110,5 @@ FROM (
             ])}
         '''
 
-        expected_output_metadata = {
-            'name': OutputMetadata(type=GraphQLString, optional=False),
-            'related_animals': OutputMetadata(
-                type=GraphQLList(GraphQLString), optional=False),
-            'related_birthdays': OutputMetadata(
-                type=GraphQLList(GraphQLDate), optional=False),
-        }
-        expected_input_metadata = {
-            'substring': GraphQLString,
-            'latest': GraphQLDate,
-        }
 
-        check_test_data(self, graphql_input, expected_match, expected_gremlin,
-                        expected_output_metadata, expected_input_metadata)
+        check_test_data(self, test_data, expected_match, expected_gremlin)
